refactor(gh): remove make_output/make_output_full re-exports from test_helpers
gh::test_helpers was re-exporting make_output and make_output_full from
crate::cmd::test_support, adding an extra indirection layer with no added
value. All other infra submodules (docker, kubectl, aws, terraform, etc.)
import these helpers directly from crate::cmd::test_support.

Remove the re-exports from gh::test_helpers so there is one canonical import
path for make_output/make_output_full across the codebase. The load_fixture
wrapper stays in test_helpers since it provides the "infra" subdir prefix,
consistent with lint::load_lint_fixture.

Update all gh sub-modules (api, issue_view, list, pr_checks, pr_view,
release_view, run_view) and mod.rs tests to import make_output/make_output_full
directly from crate::cmd::test_support.
