+++ crates/rskim/src/cmd/infra/gh/api.rs
@@ -415,8 +415,8 @@ fn base64_decoded_len(b64: &str) -> usize {
 
 #[cfg(test)]
 mod tests {
-    use super::super::test_helpers::make_output;
     use super::*;
+    use crate::cmd::test_support::make_output;
 
     // ---- JSON object ----
