+++ crates/rskim/src/cmd/infra/gh/list.rs
@@ -216,8 +216,9 @@ pub(super) fn try_parse_regex(text: &str) -> Option<InfraResult> {
 
 #[cfg(test)]
 mod tests {
-    use super::super::test_helpers::{load_fixture, make_output};
+    use super::super::test_helpers::load_fixture;
     use super::*;
+    use crate::cmd::test_support::make_output;
 
     #[test]
     fn test_tier1_gh_pass() {
