+++ crates/rskim/src/cmd/infra/gh/run_view.rs
@@ -237,8 +237,9 @@ fn try_parse_text(text: &str) -> Option<InfraResult> {
 
 #[cfg(test)]
 mod tests {
-    use super::super::test_helpers::{load_fixture, make_output};
+    use super::super::test_helpers::load_fixture;
     use super::*;
+    use crate::cmd::test_support::make_output;
 
     #[test]
     fn test_tier1_json() {
